last time

+++ src/main/java/bgu/spl/mics/application/objects/StampedDetectedObjects.java
@@ -18,7 +18,7 @@ public class StampedDetectedObjects {
         this.detectedObjects = new ArrayList<>(detectedObjects);
     }
 
-    public long getTimestamp() {
+    public int getTimestamp() {
         return time;
     }
